feat(vscode): implement symbol listing and filtering functionality
